Update by Hortense n°109

+++ src/model/particulier/Profil.java
@@ -4,14 +4,14 @@ public class Profil
 {
     private String prenom;
     private String nom;
-    private int numero;
+    private double numero;
     private String rue;
     private String id;
     private String mdp;
     private boolean estConnecte;
 
     // CONSTRUCTEUR
-    public Profil(String prenom, String nom, int numero, String rue, String id, String mdp)
+    public Profil(String prenom, String nom, double numero, String rue, String id, String mdp)
     {
         this.prenom = prenom;
         this.nom = nom;
@@ -24,7 +24,7 @@ public class Profil
     // -------- GETTERS --------
     public String getPrenom() { return prenom; }
     public String getNom() { return nom; }
-    public int getNumero() { return numero; }
+    public double getNumero() { return numero; }
     public String getRue() { return rue; }
     public String getId() { return id; }
     public String getMdp() { return mdp; }
